cross-platform support by removing some
fixed an os.name if statement changing the name of the job for meshlab to run if it isnt running on windows. The change to the name of the job (filter) for meshlab was resulting in an error on non-windows machines in the meshlab process, as the name for the job is the same across platforms. That may have not been the case in the past, or with the specific versions the author was using of meshlab.